refactor: Separate MainView navigation into individual files (#52)
- Create `TabNavHost.kt`, `ProductDetailNavHost.kt`, `ChatroomNavHost.kt`, `CreatePostNavHost.kt` to manage their respective navigation graphs.
- Move `BottomNavItem` to `presentation.main` package.
- Relocate `NavigationRoutes.kt` and `NavigationUtils.kt` to the `navigation` package.
- Update import paths in affected files to reflect these changes.
