+++ app/src/main/java/com/example/rentit/navigation/NavigationRoutes.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common.navigation
+package com.example.rentit.navigation
 
 object NavigationRoutes {
